Use 64-bit counters for amplifier keyspace
Hashcat now uses 64-bit counters for the amplifier keyspace. This lifts
the previous 32-bit restriction, allowing larger keyspace counts. All
modes are affected, but the primary impact is on the combinator and both
hybrid attacks. The _a0 and _a3 kernels are also affected, although this
is mainly theoretical since users would typically run out of memory
before loading more than 2^32 rules. If they could, it would now be
supported.

To prevent errors, hashcat checks at the start of each new attack that
the product of the base word count and the amplifier count does not
overflow an unsigned 64-bit integer. In some cases this check is
delayed, such as when using mask files, increment mode, or folders with
wordlists. This limit is still enforced and will abort sessions with the
error:

Integer overflow detected in keyspace of wordlist

This update was simpler than expected because the 64-bit amplifier index
offsets never reach the final kernels in any mode. They are either used
by a third kernel (for example, the amplifier kernel for slow hashes)
where the offset type has been updated to 64-bit, or by the host program
to offset buffers copied to constant buffers in fast hash mode.

The 32-bit nested loop iterator for fast hashes, which I initially
thought would need upgrading to 64-bit, remains unchanged. Its maximum
value is not the amplifier count but the -u value, which is capped at
1024 in hashcat.

The kernel_param_t structure was adjusted to retain status and terminal
view values. The size increase has no observed or expected performance
impact.

This is a significant change and will benefit from thorough community
testing. Basic tests indicate it works as intended, and theoretical
analysis suggests it is safe. However, edge cases may still exist and
can only be discovered through wider use.

+++ OpenCL/inc_types.h
@@ -1991,9 +1991,9 @@ typedef struct kernel_param
   u32 bitmap_shift1;        // 25
   u32 bitmap_shift2;        // 26
   u32 salt_pos_host;        // 27
-  u32 loop_pos;             // 28
-  u32 loop_cnt;             // 29
-  u32 il_cnt;               // 30
+  u64 loop_pos;             // 28
+  u64 loop_cnt;             // 29
+  u64 il_cnt;               // 30
   u32 digests_cnt;          // 31
   u32 digests_offset_host;  // 32
   u32 combs_mode;           // 33
@@ -2080,7 +2080,7 @@ typedef struct bs_word
 typedef struct plain
 {
   u64  gidvid;
-  u32  il_pos;
+  u64  il_pos;
   u32  salt_pos;
   u32  digest_pos;
   u32  hash_pos;
